finish remaining todo
