+++ tools/src/main/scala/com/nvidia/spark/rapids/tool/profiling/AutoTuner.scala
@@ -16,92 +16,173 @@
 
 package com.nvidia.spark.rapids.tool.profiling
 
-import java.io.FileNotFoundException
+import java.io.{BufferedReader, InputStreamReader}
+import java.util
 
-import scala.collection.JavaConverters._
+import scala.beans.BeanProperty
+import scala.collection.JavaConverters.mapAsScalaMapConverter
+import scala.collection.mutable
+import scala.collection.mutable.ListBuffer
 
 import org.apache.hadoop.conf.Configuration
 import org.apache.hadoop.fs.{FileSystem, FSDataInputStream, Path}
 import org.yaml.snakeyaml.Yaml
+import org.yaml.snakeyaml.constructor.{Constructor, ConstructorException}
+import org.yaml.snakeyaml.representer.Representer
 
 import org.apache.spark.internal.Logging
 
 /**
- * A wrapper class that stores all the properties that would be recommended by the Auto-tuner.
- * Separate getter and setter methods are specified for each property for ease of access.
+ * A wrapper class that stores all the GPU properties.
+ * The BeanProperty enables loading and parsing the YAML formatted content using the
+ * Constructor SnakeYaml approach.
  */
-class Config {
-  private val properties: collection.mutable.Map[String, String] = collection.mutable.Map(
-    "spark.executor.instances" -> null,
-    "spark.rapids.sql.enabled" -> null,
-    "spark.executor.cores" -> null,
-    "spark.executor.memory" -> null,
-    "spark.rapids.sql.concurrentGpuTasks" -> null,
-    "spark.task.resource.gpu.amount" -> null,
-    "spark.sql.shuffle.partitions" -> null,
-    "spark.sql.files.maxPartitionBytes" -> null,
-    "spark.rapids.memory.pinnedPool.size" -> null,
-    "spark.executor.memoryOverhead" -> null
-  )
+class GpuWorkerProps(
+    @BeanProperty var memory: String,
+    @BeanProperty var count: Int,
+    @BeanProperty var name: String) {
+  def this() {
+    this("0b", 0, "None")
+  }
+  def isMissingInfo: Boolean = {
+    count == 0 || memory.startsWith("0") || name == "None"
+  }
+  def isEmpty: Boolean = {
+    count == 0 && memory.startsWith("0") && name == "None"
+  }
+  /**
+   * If the GPU count is missing, it will set a default value based on the number of cores in the
+   * system. Adds GPU count for each 16 CPU core.
+   *
+   * @return true if the value has been updated.
+   */
+  def setDefaultGpuCountIfMissing(clusterSystemProps: SystemClusterProps): Boolean = {
+    if (count == 0) {
+      count = Math.max(1, clusterSystemProps.numCores / AutoTuner.DEF_WORKER_CPU_SORES_PER_GPU)
+      true
+    } else {
+      false
+    }
+  }
+  override def toString: String =
+    s"{Count: $count, Memory: $memory, GpuDevice: $name}"
+}
 
-  def getExecutorInstances: Int = properties("spark.executor.instances").toInt
-
-  def getExecutorCores: Int = properties("spark.executor.cores").toInt
-
-  def getExecutorMemory: String = properties("spark.executor.memory")
-
-  def getConcurrentGpuTasks: Int = properties("spark.rapids.sql.concurrentGpuTasks").toInt
-
-  def getTaskResourceGpu: Double = properties("spark.task.resource.gpu.amount").toDouble
-
-  def getShufflePartitions: Int = properties("spark.sql.shuffle.partitions").toInt
-
-  def getMaxPartitionBytes: String = properties("spark.sql.files.maxPartitionBytes")
-
-  def getPinnedPoolSize: String = properties("spark.rapids.memory.pinnedPool.size")
-
-  def getExecutorMemoryOverhead: String = properties("spark.executor.memoryOverhead")
-
-  def getAllProperties: Map[String, String] = properties.toMap
-
-  def setExecutorInstances(numInstances: Int): Unit = {
-    properties("spark.executor.instances") = numInstances.toString
+/**
+ * A wrapper class that stores all the system properties.
+ * The BeanProperty enables loading and parsing the YAML formatted content using the
+ * Constructor SnakeYaml approach.
+ */
+class SystemClusterProps(
+    @BeanProperty var numCores: Int,
+    @BeanProperty var memory: String,
+    @BeanProperty var numWorkers: Int) {
+  def this() {
+    this(0, "0b", 0)
+  }
+  def isMissingInfo: Boolean = {
+    numCores == 0 || memory.startsWith("0") || numWorkers == 0
+  }
+  def isEmpty: Boolean = {
+    numCores == 0 && memory.startsWith("0") && numWorkers == 0
   }
 
-  def setExecutorCores(executorCores: Int): Unit = {
-    properties("spark.executor.cores") = executorCores.toString
+  /**
+   * Set the number of cores to default if it was 0.
+   *
+   * @return true if the value has been updated.
+   */
+  def setDefaultNumCoresIfMissing(coreNum: Int): Boolean = {
+    if (numCores == 0) {
+      numCores = coreNum
+      true
+    } else {
+      false
+    }
+  }
+  override def toString: String =
+    s"{numCores: $numCores, Memory: $memory, numWorkers: $numWorkers}"
+}
+
+/**
+ * A wrapper class that stores all the properties of the cluster.
+ * The BeanProperty enables loading and parsing the YAML formatted content using the
+ * Constructor SnakeYaml approach.
+ *
+ * @param system wrapper that includes the properties related to number of cores and memory.
+ * @param gpu wrapper that includes the properties related to GPU.
+ * @param softwareProperties a set of software properties such as Spark properties.
+ *                           The properties are typically loaded from the default cluster
+ *                           configurations.
+ */
+class ClusterProperties(
+    @BeanProperty var system: SystemClusterProps,
+    @BeanProperty var gpu: GpuWorkerProps,
+    @BeanProperty var softwareProperties: util.LinkedHashMap[String, String]) {
+
+  import AutoTuner._
+
+  def this() {
+    this(new SystemClusterProps(), new GpuWorkerProps(), new util.LinkedHashMap[String, String]())
+  }
+  def isEmpty: Boolean = {
+    system.isEmpty && gpu.isEmpty
+  }
+  def getTargetProperties: mutable.Map[String, String] = {
+    softwareProperties.asScala.filter(entry => recommendationsTarget.contains(entry._1))
+  }
+  override def toString: String =
+    s"{${system.toString}, ${gpu.toString}, $softwareProperties}"
+}
+
+/**
+ * Wrapper to hold the recommendation of a given criterion.
+ *
+ * @param name the property label.
+ * @param originalValue the value loaded from the spark properties.
+ * @param recommended the recommended value by the AutoTuner.
+ */
+class RecommendationEntry(val name: String, val originalValue: String, var recommended: String) {
+  def setRecommendedValue(value: String): Unit = {
+    if (value != null) {
+      recommended = value
+    }
   }
 
-  def setExecutorMemory(executorMemory: String): Unit = {
-    properties("spark.executor.memory") = executorMemory
+  private def getRawValue(arg: String): String = {
+    if (arg != null && AutoTuner.containsMemoryUnits(arg)) {
+      // if it is memory return the bytesUnit
+      s"${AutoTuner.convertFromHumanReadableSize(arg)}"
+    } else {
+      arg
+    }
   }
 
-  def setConcurrentGpuTasks(concurrentGpuTasks: Int): Unit = {
-    properties("spark.rapids.sql.concurrentGpuTasks") = concurrentGpuTasks.toString
+  /**
+   * Returns true when the recommendation is different than the original.
+   */
+  private def recommendsNewValue(): Boolean = {
+    val convertedOriginal = getRawValue(originalValue)
+    val convertedRecommended = getRawValue(recommended)
+    convertedOriginal != convertedRecommended
   }
 
-  def setTaskResourceGpu(taskResourceGpu: Double): Unit = {
-    properties("spark.task.resource.gpu.amount") = taskResourceGpu.toString
-  }
-
-  def setShufflePartitions(shufflePartitions: Int): Unit = {
-    properties("spark.sql.shuffle.partitions") = shufflePartitions.toString
-  }
-
-  def setMaxPartitionBytes(maxPartitionBytes: String): Unit = {
-    properties("spark.sql.files.maxPartitionBytes") = maxPartitionBytes
-  }
-
-  def setPinnedPoolSize(pinnedPoolSize: String): Unit = {
-    properties("spark.rapids.memory.pinnedPool.size") = pinnedPoolSize
-  }
-
-  def setExecutorMemoryOverhead(executorMemoryOverhead: String): Unit = {
-    properties("spark.executor.memoryOverhead") = executorMemoryOverhead
-  }
-
-  def setExecutorMemoryOverheadFactor(executorMemoryOverheadFactor: Double): Unit = {
-    properties("spark.executor.memoryOverheadFactor") = executorMemoryOverheadFactor.toString
+  /**
+   * True or False whether the recommendation is valid. e.g., recommendations that does not change
+   * the original value returns false if filter is enabled.
+   * @param filterByUpdated flag to pick only the properties that would be updated by the
+   *                        recommendations
+   */
+  def isValid(filterByUpdated: Boolean): Boolean = {
+    if (recommended == null) { // no recommendations
+      false
+    } else {
+      if (filterByUpdated) { // filter enabled
+        recommendsNewValue()
+      } else {
+        true
+      }
+    }
   }
 }
 
@@ -109,7 +190,7 @@ class Config {
  * AutoTuner module that uses event logs and worker's system properties to recommend Spark
  * RAPIDS configuration based on heuristics.
  *
- * Example (Refer to test suite for more cases):
+ * Example:
  * a. Success:
  *    Input:
  *      system:
@@ -123,212 +204,364 @@ class Config {
  *        count: 8
  *        memory: 32gb
  *        name: NVIDIA V100
+ *      softwareProperties:
+ *        spark.driver.maxResultSize: 7680m
+ *        spark.driver.memory: 15360m
+ *        spark.executor.cores: '8'
+ *        spark.executor.instances: '2'
+ *        spark.executor.memory: 47222m
+ *        spark.executorEnv.OPENBLAS_NUM_THREADS: '1'
+ *        spark.extraListeners: com.google.cloud.spark.performance.DataprocMetricsListener
+ *        spark.scheduler.mode: FAIR
+ *        spark.sql.cbo.enabled: 'true'
+ *        spark.ui.port: '0'
+ *        spark.yarn.am.memory: 640m
  *
  *    Output:
  *       Spark Properties:
  *       --conf spark.executor.cores=8
- *       --conf spark.executor.instances=32
- *       --conf spark.executor.memory=63.75g
- *       --conf spark.executor.memoryOverhead=8.38g
- *       --conf spark.rapids.memory.pinnedPool.size=2g
- *       --conf spark.rapids.sql.concurrentGpuTasks=4
- *       --conf spark.sql.files.maxPartitionBytes=4g
- *       --conf spark.sql.shuffle.partitions=200
+ *       --conf spark.executor.instances=20
+ *       --conf spark.executor.memory=16384m
+ *       --conf spark.executor.memoryOverhead=5734m
+ *       --conf spark.rapids.memory.pinnedPool.size=4096m
+ *       --conf spark.rapids.sql.concurrentGpuTasks=2
+ *       --conf spark.sql.files.maxPartitionBytes=4096m
  *       --conf spark.task.resource.gpu.amount=0.125
  *
+ *       Comments:
+ *       - 'spark.rapids.sql.concurrentGpuTasks' was not set.
+ *       - 'spark.executor.memoryOverhead' was not set.
+ *       - 'spark.rapids.memory.pinnedPool.size' was not set.
+ *       - 'spark.sql.adaptive.enabled' should be enabled for better performance.
+ *
  * b. Failure:
- *    Input: Empty or Incorrect System Properties
+ *    Input: Incorrect File
  *    Output:
+ *      Cannot recommend properties. See Comments.
+ *
  *      Comments:
+ *      - java.io.FileNotFoundException: File worker_info.yaml does not exist
  *      - 'spark.executor.memory' should be set to at least 2GB/core.
- *      - 'spark.executor.instances' should be set to 'num_gpus * num_workers'.
- *      - 'spark.task.resource.gpu.amount' should be set to 1/#cores.
+ *      - 'spark.executor.instances' should be set to (gpuCount * numWorkers).
+ *      - 'spark.task.resource.gpu.amount' should be set to Max(1, (numCores / gpuCount)).
+ *      - 'spark.rapids.sql.concurrentGpuTasks' should be set between [1, 4]
+ *      - 'spark.rapids.memory.pinnedPool.size' should be set up to 4096m.
+ *      - 'spark.sql.adaptive.enabled' should be enabled for better performance.
+ *
+ * @param clusterProps The cluster properties including cores, mem, GPU, and software
+ *                     (see [[ClusterProperties]]).
+ * @param appInfo the container holding the profiling result.
  */
-class AutoTuner(app: ApplicationSummaryInfo, workerInfo: String) extends Logging {
+class AutoTuner(
+    val clusterProps: ClusterProperties,
+    val appInfo: Option[ApplicationSummaryInfo])  extends Logging {
+
   import AutoTuner._
-  val DEFAULT_SHUFFLE_PARTITION_MULTIPLIER: Int = 2
-  val MAX_JVM_GCTIME_FRACTION: Double = 0.3
 
-  val DEFAULT_CONCURRENT_GPU_TASKS: Int = 2
-  val DEFAULT_CONCURRENT_GPU_TASKS_MULTIPLIER: Double = 0.125 // Currently aggressively set to 1/8
-  val MAX_CONCURRENT_GPU_TASKS: Int = 4
+  var comments = new ListBuffer[String]()
+  var recommendations: mutable.LinkedHashMap[String, RecommendationEntry] =
+    mutable.LinkedHashMap[String, RecommendationEntry]()
+  // list of criterion to be computed for recommendations
+  private var skipRecommendations: Option[Seq[String]] = None
+  // When enabled, the profiler recommendations should only include updated settings.
+  var filterByUpdatedPropertiesEnabled: Boolean = true
 
-  val DEFAULT_MAX_PARTITION_BYTES: String = "512m"
-  val MAX_PARTITION_BYTES_BOUND: String = "4g"
-  val MAX_PARTITION_BYTES_RANGE: String = "256m"
-  val MIN_PARTITION_BYTES_RANGE: String = "128m"
+  def getSparkPropertyFromProfile(propKey: String): Option[String] = {
+    if (appInfo.isDefined) {
+      appInfo.get.sparkProps.collectFirst {
+        case entry: RapidsPropertyProfileResult
+          if entry.key == propKey && entry.rows(1) != "null" => entry.rows(1)
+      }
+    } else {
+      None
+    }
+  }
 
-  val DEFAULT_PINNED_POOL_SIZE: String = "2g"
-  val DEFAULT_MEMORY_OVERHEAD_FACTOR: Double = 0.1
-  val DEFAULT_SYSTEM_OVERHEAD: String = "2g" // Overhead of other system processes
+  def getPropertyValue(key: String): Option[String] = {
+    val fromProfile = getSparkPropertyFromProfile(key)
+    if (fromProfile.isDefined) {
+      fromProfile
+    } else {
+      Option(clusterProps.softwareProperties.get(key))
+    }
+  }
 
-  val MAX_PER_EXECUTOR_CORE_COUNT: Int = 16
-  val MIN_PER_EXECUTOR_CORE_COUNT: Int = 4
+  def initRecommendations(): Unit = {
+    recommendationsTarget.foreach { key =>
+      getPropertyValue(key).foreach { propVal =>
+        val recommendationVal = new RecommendationEntry(key, propVal, null)
+        recommendations(key) = recommendationVal
+      }
+    }
+  }
 
-  val MAX_EXECUTOR_MEMORY: String = "64g"
-  val MIN_EXECUTOR_MEMORY: String = "8g"
+  def appendRecommendation(key: String, value: String): Unit = {
+    val skip = skipRecommendations match {
+      case Some(seq) => seq.contains(key)
+      case None => false
+    }
+    if (!skip) {
+      val recomRecord = recommendations.getOrElseUpdate(key,
+        new RecommendationEntry(key, null, null))
+      recomRecord.setRecommendedValue(value)
+      if (recomRecord.originalValue == null && value != null) {
+        // add a comment
+        appendComment(s"'$key' was not set.")
+      }
+    }
+  }
 
-  var comments: Seq[String] = Seq()
+  /**
+   * Safely appends the recommendation to the given key.
+   * It skips if the value is 0.
+   */
+  def appendRecommendationForIntNum(key: String, value: Int): Unit = {
+    if (value != 0) {
+      appendRecommendation(key: String, s"$value")
+    }
+  }
+
+  /**
+   * Safely appends the recommendation to the given key.
+   * It skips if the value is 0.0.
+   */
+  def appendRecommendationForFloat(key: String, value: Double): Unit = {
+    if (value != 0.0) {
+      appendRecommendation(key: String, s"$value")
+    }
+  }
+  /**
+   * Safely appends the recommendation to the given key.
+   * It appends "m" to the string value. It skips if the value is 0 or null.
+   */
+  def appendRecommendationForMemoryMB(key: String, value: String): Unit = {
+    if (value != null && value.toDouble != 0.0) {
+      appendRecommendation(key, s"${value}m")
+    }
+  }
 
   /**
    * Recommendation for 'spark.executor.instances' based on number of gpus and workers.
    * Assumption - In case GPU properties are not available, it assumes 1 GPU per worker.
    */
-  private def recommendExecutorInstances(recommendedConfig: Config,
-      systemProps: SystemProps): Unit = {
-    systemProps.numWorkers match {
-      case Some(numWorkers) =>
-        val numInstances = if (systemProps.gpuProps != null) {
-          numWorkers * systemProps.gpuProps.count
-        } else {
-          // Assumption: 1 GPU per worker.
-          numWorkers
-        }
-
-        recommendedConfig.setExecutorInstances(numInstances)
-      case None =>
-        val num_gpus_str = if (systemProps.gpuProps != null) {
-          systemProps.gpuProps.count.toString
-        } else {
-          "num_gpus"
-        }
-
-        comments :+= s"'spark.executor.instances' should be set to $num_gpus_str * num_workers."
-    }
+  def calcExecInstances(): Int = {
+    clusterProps.gpu.getCount * clusterProps.system.numWorkers
   }
 
   /**
    * Recommendation for 'spark.executor.cores' based on number of cpu cores and gpus.
    */
-  private def recommendExecutorCores(recommendedConfig: Config,
-      systemProps: SystemProps): Unit = {
-    val numCores: Int = if (systemProps.gpuProps != null) {
-      Math.min(systemProps.numCores * 1.0 / systemProps.gpuProps.count,
-        MAX_PER_EXECUTOR_CORE_COUNT).toInt
+  def calcNumExecutorCores: Int = {
+    val executorsPerNode = clusterProps.gpu.getCount
+    if (executorsPerNode == 0) {
+      // Being extra caution about zero division.
+      // handling 0 and default value is processed in [[processPropsAndCheck]]
+      1
     } else {
-      systemProps.numCores
+      Math.max(1, clusterProps.system.getNumCores / executorsPerNode)
     }
+  }
 
-    if (numCores < MIN_PER_EXECUTOR_CORE_COUNT) {
-      comments :+= s"Number of cores per executor is very low. " +
-        s"It is recommended to have at least $MIN_PER_EXECUTOR_CORE_COUNT cores per executor."
+  /**
+   * Recommendation for 'spark.task.resource.gpu.amount' based on num of cpu cores.
+   */
+  def calcTaskGPUAmount(numExecCoresCalculator: () => Int): Double = {
+    val numExecutorCores =  numExecCoresCalculator.apply()
+    if (numExecutorCores == 0) 0.0 else 1.0 / numExecutorCores
+  }
+
+  /**
+   * Recommendation for 'spark.rapids.sql.concurrentGpuTasks' based on gpu memory.
+   */
+  def calcGpuConcTasks(): Long = {
+    Math.min(MAX_CONC_GPU_TASKS,
+      convertToMB(clusterProps.gpu.memory) / DEF_GPU_MEM_PER_TASK_MB)
+  }
+
+  /**
+   * Calculates the available memory for each executor on the worker based on the number of
+   * executors per node and the memory.
+   */
+  private def calcAvailableMemPerExec(): Double = {
+    // account for system overhead
+    val usableWorkerMem =
+      Math.max(0, convertToMB(clusterProps.system.memory) - DEF_SYSTEM_RESERVE_MB)
+    val executorsPerNode = clusterProps.gpu.getCount
+    if (executorsPerNode == 0) {
+      0.0
+    } else {
+      (1.0 * usableWorkerMem) / executorsPerNode
     }
-
-    if (systemProps.numWorkers.nonEmpty) {
-      val numInstances = recommendedConfig.getExecutorInstances
-      if (numCores * numInstances < systemProps.numCores) {
-        comments :+= "Not all cores in the machine are being used. " +
-          "It is recommended to use different machine."
-      }
-    }
-
-    recommendedConfig.setExecutorCores(numCores)
   }
 
   /**
    * Recommendation for 'spark.executor.memory' based on system memory, cluster scheduler
    * and num of gpus.
-   * Note - This method requires recommendedConfig to have 'spark.executor.cores' set.
    */
-  private def recommendExecutorMemory(recommendedConfig: Config,
-      systemProps: SystemProps): Unit = {
-    val systemMemoryNum: Long = convertFromHumanReadableSize(systemProps.memory)
-    val systemOverhead: Long = convertFromHumanReadableSize(DEFAULT_SYSTEM_OVERHEAD)
-    val effectiveSystemMemoryNum: Long =  systemMemoryNum - systemOverhead -
-          convertFromHumanReadableSize(DEFAULT_PINNED_POOL_SIZE)
-    val maxExecutorMemNum: Long = convertFromHumanReadableSize(MAX_EXECUTOR_MEMORY)
-
-    val executorMemory: Long = if (systemProps.gpuProps != null) {
-      Math.min(effectiveSystemMemoryNum * 1.0 / systemProps.gpuProps.count,
-        maxExecutorMemNum).toLong
-    } else {
-      Math.min(effectiveSystemMemoryNum * 1.0 / recommendedConfig.getExecutorCores,
-        maxExecutorMemNum).toLong
-    }
-
-    if(executorMemory < convertFromHumanReadableSize(MIN_EXECUTOR_MEMORY)) {
-      comments :+= s"Executor memory is very low. " +
-        s"It is recommended to have at least $MIN_EXECUTOR_MEMORY."
-    }
-
-    recommendedConfig.setExecutorMemory(convertToHumanReadableSize(executorMemory))
+  def calcExecutorHeap(executorContainerMemCalculator: () => Double,
+      numExecCoresCalculator: () => Int): Long = {
+    // reserve 10% of heap as memory overhead
+    val maxExecutorHeap = Math.max(0,
+      executorContainerMemCalculator() * (1 - DEF_HEAP_OVERHEAD_FRACTION)).toInt
+    // give up to 2GB of heap to each executor core
+    Math.min(maxExecutorHeap, DEF_HEAP_PER_CORE_MB * numExecCoresCalculator())
   }
 
   /**
-   * Calculate memory overhead as: pinnedPoolSize + (memoryOverheadFactor * executorMemory)
+   * Recommendation for 'spark.rapids.memory.pinnedPool.size' and 'spark.executor.memoryOverhead'
+   * based on executor memory.
    */
-  private def calculateMemoryOverhead(pinnedPoolSize: String, executorMemory: String): Long = {
-    val pinnedPoolSizeNum = convertFromHumanReadableSize(pinnedPoolSize)
-    val executorMemoryNum = convertFromHumanReadableSize(executorMemory)
-    (pinnedPoolSizeNum + DEFAULT_MEMORY_OVERHEAD_FACTOR * executorMemoryNum).toLong
+  def calcPinnedMemoryWithOverhead(
+      execHeapCalculator: () => Long,
+      containerMemCalculator: () => Double): (Long, Long) = {
+    val executorHeap = execHeapCalculator()
+    var executorMemOverhead = (executorHeap * DEF_HEAP_OVERHEAD_FRACTION).toLong
+    // pinned memory uses any unused space up to 4GB
+    val pinnedMem = Math.min(MAX_PINNED_MEMORY_MB,
+      containerMemCalculator.apply() - executorHeap - executorMemOverhead).toLong
+    executorMemOverhead += pinnedMem
+    (pinnedMem, executorMemOverhead)
+  }
+
+  private def getSparkVersion: Option[String] = {
+    appInfo match {
+      case Some(app) => Option(app.appInfo.head.sparkVersion)
+      case None => None
+    }
   }
 
   /**
-   * Recommendation for 'spark.rapids.memory.pinnedPool.size' if it is not set.
-   * Recommendation for memoryOverhead and memoryOverheadFactor based on cluster scheduler and
-   * spark version used.
-   *
-   * Flow:
-   * if (pinnedPoolSize is not set) -> set pinnedPoolSize, calculate and set memoryOverhead
-   * else
-   *   if using yarn
-   *     -> if memoryOverhead is not set, calculate and set.
-   *   if using k8s
-   *     -> if memoryOverheadFactor is not set, add comment.
+   * Find the label of the memory.overhead based on the spark master configuration and the spark
+   * version.
+   * @return "spark.executor.memoryOverhead", "spark.kubernetes.memoryOverheadFactor",
+   *         or "spark.executor.memoryOverheadFactor".
    */
-  private def recommendMemoryOverheadProperties(recommendedConfig: Config): Unit = {
-    val pinnedPoolSize = getSparkProperty(app, "spark.rapids.memory.pinnedPool.size")
-    if (pinnedPoolSize.isEmpty) {
-      recommendedConfig.setPinnedPoolSize(DEFAULT_PINNED_POOL_SIZE)
-      val memoryOverhead = calculateMemoryOverhead(DEFAULT_PINNED_POOL_SIZE,
-        recommendedConfig.getExecutorMemory)
-      recommendedConfig.setExecutorMemoryOverhead(convertToHumanReadableSize(memoryOverhead))
-    } else {
-      val sparkMaster = getSparkProperty(app, "spark.master")
-      if (sparkMaster.contains("yarn")) {
-        if (getSparkProperty(app, "spark.executor.memoryOverhead").isEmpty) {
-          val memoryOverhead = calculateMemoryOverhead(pinnedPoolSize.get,
-            recommendedConfig.getExecutorMemory)
-          recommendedConfig.setExecutorMemoryOverhead(
-            convertToHumanReadableSize(memoryOverhead))
-        }
-      } else if (sparkMaster.contains("k8s")) {
-        if (compareSparkVersion(app.appInfo.head.sparkVersion, "3.3.0") > 0) {
-          if (getSparkProperty(app, "spark.executor.memoryOverheadFactor").isEmpty) {
-            comments :+= "'spark.executor.memoryOverheadFactor' must be set " +
-              "if using 'spark.rapids.memory.pinnedPool.size'"
+  def memoryOverheadLabel: String = {
+    val sparkMasterConf = getPropertyValue("spark.master")
+    val defaultLabel = "spark.executor.memoryOverhead"
+    sparkMasterConf match {
+      case None => defaultLabel
+      case Some(sparkMaster) =>
+        if (sparkMaster.contains("yarn")) {
+          defaultLabel
+        } else if (sparkMaster.contains("k8s")) {
+          getSparkVersion match {
+            case Some(version) =>
+              if (compareSparkVersion(version, "3.3.0") > 0) {
+                "spark.executor.memoryOverheadFactor"
+              } else {
+                "spark.kubernetes.memoryOverheadFactor"
+              }
+            case None => defaultLabel
           }
         } else {
-          if (getSparkProperty(app, "spark.kubernetes.memoryOverheadFactor").isEmpty) {
-            comments :+= "'spark.kubernetes.memoryOverheadFactor' must be set " +
-              "if using 'spark.rapids.memory.pinnedPool.size'"
+          defaultLabel
+        }
+    }
+  }
+
+  /**
+   * Flow:
+   *   if "spark.rapids.memory.pinnedPool.size" is set
+   *     if yarn -> recommend "spark.executor.memoryOverhead"
+   *     if using k8s ->
+   *         if version > 3.3.0 recommend "spark.executor.memoryOverheadFactor" and add comment
+   *         else recommend "spark.kubernetes.memoryOverheadFactor" and add comment if missing
+   */
+  def addRecommendationForMemoryOverhead(recomValue: String): Unit = {
+    val memOverheadLookup = memoryOverheadLabel
+    appendRecommendationForMemoryMB(memOverheadLookup, recomValue)
+    getPropertyValue("spark.rapids.memory.pinnedPool.size").foreach { lookup =>
+      if (lookup != "spark.executor.memoryOverhead") {
+        if (getPropertyValue(memOverheadLookup).isEmpty) {
+          appendComment(s"'$memOverheadLookup' must be set if using " +
+            s"'spark.rapids.memory.pinnedPool.size")
+        }
+      }
+    }
+  }
+
+  def calculateRecommendations(): Unit = {
+    appendRecommendationForIntNum("spark.executor.instances", calcExecInstances())
+    val numExecutorCores = calcNumExecutorCores
+    val execCoresExpr = () => numExecutorCores
+
+    appendRecommendationForIntNum("spark.executor.cores", numExecutorCores)
+    appendRecommendationForFloat("spark.task.resource.gpu.amount",
+      calcTaskGPUAmount(execCoresExpr))
+    appendRecommendationForIntNum("spark.rapids.sql.concurrentGpuTasks",
+      calcGpuConcTasks().toInt)
+    val availableMemPerExec = calcAvailableMemPerExec()
+    val availableMemPerExecExpr = () => availableMemPerExec
+    val executorHeap = calcExecutorHeap(availableMemPerExecExpr, execCoresExpr)
+    val executorHeapExpr = () => executorHeap
+    appendRecommendationForMemoryMB("spark.executor.memory", s"$executorHeap")
+    val (pinnedMemory, memoryOverhead) =
+      calcPinnedMemoryWithOverhead(executorHeapExpr, availableMemPerExecExpr)
+    appendRecommendationForMemoryMB("spark.rapids.memory.pinnedPool.size", s"$pinnedMemory")
+    addRecommendationForMemoryOverhead(s"$memoryOverhead")
+
+    recommendMaxPartitionBytes()
+    recommendShufflePartitions()
+    recommendGeneralProperties()
+  }
+
+  /**
+   * Checks whether the cluster properties are valid.
+   * If the cluster worker-info is missing entries (i.e., CPU and GPU count), it sets the entries
+   * to default values. For each default value, a comment is added to the [[comments]].
+   *
+   * @return false if the cluster properties are not loaded. e.g, all entries are set to 0.
+   *         true if the missing information were updated to default initial values.
+   */
+  def processPropsAndCheck: Boolean = {
+    if (clusterProps.isEmpty) {
+      false
+    } else {
+      if (clusterProps.system.isMissingInfo) {
+        // set the cores to DEF_WORKER_CPU_CORES by default
+        val systemComment =
+          if (clusterProps.system.setDefaultNumCoresIfMissing(DEF_WORKER_CPU_CORES)) {
+            s"CPU cores is missing. Setting default to ${clusterProps.system.getNumCores}."
+          } else {
+            s"CPU properties is incomplete: ${clusterProps.system}."
           }
+        appendComment(systemComment)
+      }
+      if (clusterProps.gpu.isMissingInfo) {
+        val gpuComment =
+          if (clusterProps.gpu.setDefaultGpuCountIfMissing(clusterProps.system)) {
+            s"GPU count is missing. Setting default to ${clusterProps.gpu.getCount}."
+          } else {
+            s"GPU properties is incomplete: ${clusterProps.gpu}."
+          }
+        appendComment(gpuComment)
+      }
+      true
+    }
+  }
+
+  private def recommendGeneralProperties(): Unit = {
+    val aqeEnabled = getPropertyValue("spark.sql.adaptive.enabled").getOrElse("False")
+    if (aqeEnabled == "False") {
+      appendComment(commentsForMissingProps("spark.sql.adaptive.enabled"))
+    }
+    if (appInfo.isDefined) {
+      val jvmGCFraction = appInfo.get.sqlTaskAggMetrics.map {
+        taskMetrics => taskMetrics.jvmGCTimeSum * 1.0 / taskMetrics.executorCpuTime
+      }
+      if (jvmGCFraction.nonEmpty) { // avoid zero division
+        if ((jvmGCFraction.sum / jvmGCFraction.size) > MAX_JVM_GCTIME_FRACTION) {
+          appendComment("Average JVM GC time is very high. " +
+            "Other Garbage Collectors can be used for better performance.")
         }
       }
     }
   }
 
   /**
-   * Recommendation for 'spark.sql.shuffle.partitions' based on spill size.
-   */
-  private def recommendShufflePartitions(recommendedConfig: Config): Unit = {
-    var shufflePartitions: Int = getSparkProperty(app, "spark.sql.shuffle.partitions")
-      .getOrElse("200").toInt
-
-    // TODO: Need to look at other metrics for GPU spills (DEBUG mode), and batch sizes metric
-    val totalSpilledMetrics = app.sqlTaskAggMetrics.map {
-      task => task.diskBytesSpilledSum + task.memoryBytesSpilledSum
-    }.sum
-    if (totalSpilledMetrics > 0) {
-      shufflePartitions *= DEFAULT_SHUFFLE_PARTITION_MULTIPLIER
-      // Could be memory instead of partitions
-      comments :+= "\"spark.sql.shuffle.partitions\" should be increased since spilling occurred."
-    }
-    recommendedConfig.setShufflePartitions(shufflePartitions)
-  }
-
-  /**
-   * Calculate max partition bytes using task input size.
+   * Calculate max partition bytes using task input size and existing setting for maxPartitionBytes.
+   * Note that this won't apply the same on iceberg.
    * Eg,
    * MIN_PARTITION_BYTES_RANGE = 128m, MAX_PARTITION_BYTES_RANGE = 256m
    * (1) Input:  maxPartitionBytes = 512m
@@ -339,157 +572,226 @@ class AutoTuner(app: ApplicationSummaryInfo, workerInfo: String) extends Logging
    *     Output: newMaxPartitionBytes = 2g / (512m/128m) = 512m
    */
   private def calculateMaxPartitionBytes(maxPartitionBytes: String): String = {
-    val taskInputSize =
-      app.sqlTaskAggMetrics.map(_.inputBytesReadAvg).sum / app.sqlTaskAggMetrics.size
-    val maxPartitionBytesNum = convertFromHumanReadableSize(maxPartitionBytes)
-    if (taskInputSize > 0 &&
-      taskInputSize < convertFromHumanReadableSize(MIN_PARTITION_BYTES_RANGE)) {
-      // Increase partition size
-      val calculatedMaxPartitionBytes = Math.min(
-        maxPartitionBytesNum *
-          (convertFromHumanReadableSize(MIN_PARTITION_BYTES_RANGE) / taskInputSize),
-        convertFromHumanReadableSize(MAX_PARTITION_BYTES_BOUND))
-
-      convertToHumanReadableSize(calculatedMaxPartitionBytes.toLong)
-    } else if (taskInputSize > convertFromHumanReadableSize(MAX_PARTITION_BYTES_RANGE)) {
-      // Decrease partition size
-      val calculatedMaxPartitionBytes = Math.min(
-        maxPartitionBytesNum /
-          (taskInputSize / convertFromHumanReadableSize(MAX_PARTITION_BYTES_RANGE)),
-        convertFromHumanReadableSize(MAX_PARTITION_BYTES_BOUND))
-
-      convertToHumanReadableSize(calculatedMaxPartitionBytes.toLong)
+    val app = appInfo.get
+    val maxPartitionBytesNum = convertToMB(maxPartitionBytes)
+    if (app.sqlTaskAggMetrics.isEmpty) { // avoid division by zero
+      maxPartitionBytesNum.toString
     } else {
-      // Do not recommend maxPartitionBytes
-      null
+      val taskInputSizeInMB =
+        (app.sqlTaskAggMetrics.map(_.inputBytesReadAvg).sum / app.sqlTaskAggMetrics.size) / 1024
+      if (taskInputSizeInMB > 0 &&
+        taskInputSizeInMB < MIN_PARTITION_BYTES_RANGE_MB) {
+        // Increase partition size
+        val calculatedMaxPartitionBytes = Math.min(
+          maxPartitionBytesNum *
+            (MIN_PARTITION_BYTES_RANGE_MB / taskInputSizeInMB),
+          MAX_PARTITION_BYTES_BOUND_MB)
+        calculatedMaxPartitionBytes.toLong.toString
+      } else if (taskInputSizeInMB > MAX_PARTITION_BYTES_RANGE_MB) {
+        // Decrease partition size
+        val calculatedMaxPartitionBytes = Math.min(
+          maxPartitionBytesNum /
+            (taskInputSizeInMB / MAX_PARTITION_BYTES_RANGE_MB),
+          MAX_PARTITION_BYTES_BOUND_MB)
+        calculatedMaxPartitionBytes.toLong.toString
+      } else {
+        // Do not recommend maxPartitionBytes
+        null
+      }
     }
   }
 
   /**
    * Recommendation for 'spark.sql.files.maxPartitionBytes' based on input size for each task.
    */
-  private def recommendMaxPartitionBytes(recommendedConfig: Config): Unit = {
-    getSparkProperty(app, "spark.sql.files.maxPartitionBytes") match {
-      case None =>
-        val newMaxPartitionBytes = calculateMaxPartitionBytes(DEFAULT_MAX_PARTITION_BYTES)
-        recommendedConfig.setMaxPartitionBytes(newMaxPartitionBytes)
+  private def recommendMaxPartitionBytes(): Unit = {
+    val res =
+      getPropertyValue("spark.sql.files.maxPartitionBytes") match {
+        case None =>
+          if (appInfo.isDefined) {
+            calculateMaxPartitionBytes(MAX_PARTITION_BYTES)
+          } else {
+            s"${convertToMB(MAX_PARTITION_BYTES)}"
+          }
+        case Some(maxPartitionBytes) =>
+          if (appInfo.isDefined) {
+            calculateMaxPartitionBytes(maxPartitionBytes)
+          } else {
+            maxPartitionBytes
+          }
+      }
+    appendRecommendationForMemoryMB("spark.sql.files.maxPartitionBytes", res)
+  }
 
-      case Some(maxPartitionBytes) =>
-        val newMaxPartitionBytes = calculateMaxPartitionBytes(maxPartitionBytes)
-        recommendedConfig.setMaxPartitionBytes(newMaxPartitionBytes)
-        comments :+= s"Although 'spark.sql.files.maxPartitionBytes' was set to $maxPartitionBytes" +
-          s", recommended value is $newMaxPartitionBytes."
+  private def recommendShufflePartitions(): Unit = {
+    val lookup = "spark.sql.shuffle.partitions"
+    var shufflePartitions =
+      getPropertyValue(lookup).getOrElse(DEF_SHUFFLE_PARTITIONS).toInt
+    // TODO: Need to look at other metrics for GPU spills (DEBUG mode), and batch sizes metric
+    if (appInfo.isDefined) {
+      val totalSpilledMetrics = appInfo.get.sqlTaskAggMetrics.map {
+        task => task.diskBytesSpilledSum + task.memoryBytesSpilledSum
+      }.sum
+      if (totalSpilledMetrics > 0) {
+        shufflePartitions *= DEF_SHUFFLE_PARTITION_MULTIPLIER
+        // Could be memory instead of partitions
+        appendOptionalComment(lookup,
+          s"'$lookup' should be increased since spilling occurred.")
+      }
+    }
+    appendRecommendation("spark.sql.shuffle.partitions", s"$shufflePartitions")
+  }
+
+  def appendOptionalComment(lookup: String, comment: String): Unit = {
+    val skip = skipRecommendations match {
+      case None => false
+      case Some(criteriaSeq) => criteriaSeq.contains(lookup)
+    }
+    if (!skip) {
+      appendComment(comment)
     }
   }
 
-  private def recommendGeneralProperties() = {
-    val aqeEnabled = getSparkProperty(app, "spark.sql.adaptive.enabled").getOrElse("False")
-    if (aqeEnabled == "False") {
-      comments :+= "'spark.sql.adaptive.enabled' should be enabled for better performance."
-    }
+  def appendComment(comment: String): Unit = {
+    comments += comment
+  }
 
-    val jvmGCFraction = app.sqlTaskAggMetrics.map {
-      taskMetrics => taskMetrics.jvmGCTimeSum * 1.0 / taskMetrics.executorCpuTime
-    }
-    if ((jvmGCFraction.sum / jvmGCFraction.size) > MAX_JVM_GCTIME_FRACTION) {
-      comments :+= "Average JVM GC time is very high. " +
-        "Other Garbage Collectors can be used for better performance"
-    }
+  def convertClusterPropsToString(): String = {
+    clusterProps.toString
+  }
+
+  private def toCommentProfileResult:
+  Seq[RecommendedCommentResult] = {
+    comments.map(RecommendedCommentResult)
+  }
+
+  private def toRecommendationsProfileResult: Seq[RecommendedPropertyResult] = {
+    val finalRecommendations =
+      recommendations.filter(elem => elem._2.isValid(filterByUpdatedPropertiesEnabled))
+    finalRecommendations.collect {
+      case (key, record) => RecommendedPropertyResult(key, record.recommended)
+    }.toSeq.sortBy(_.property)
   }
 
   /**
-   * Recommendation for 'spark.task.resource.gpu.amount' based on num of cpu cores.
+   * The Autotuner loads the spark properties from either the ClusterProperties or the eventlog.
+   * 1- runs the calculation for each criterion and saves it as a [[RecommendationEntry]].
+   * 2- The final list of recommendations include any [[RecommendationEntry]] that has a
+   *    recommendation that is different from the original property.
+   * 3- Null values are excluded.
+   * 4- A comment is added for each missing property in the spark property.
+   *
+   * @param skipList a list of recommendations to be skipped. If none, all recommendations are
+   *                 returned. Default is None.
+   * @param showOnlyUpdatedProps When enabled, the profiler recommendations should only include
+   *                             updated settings.
    */
-  private def recommendTaskResourceGpu(recommendedConfig: Config): Unit = {
-    val numCores: Int = recommendedConfig.getExecutorCores
-    val taskResourceGpu: Double = 1.0 / numCores
-    recommendedConfig.setTaskResourceGpu(taskResourceGpu)
-  }
-
-  /**
-   * Recommendation for 'spark.rapids.sql.concurrentGpuTasks' based on gpu memory.
-   */
-  private def recommendConcurrentGpuTasks(recommendedConfig: Config,
-      systemProps: SystemProps): Unit = {
-    val numCores: Int = recommendedConfig.getExecutorCores
-    val gpuMemoryNum: Long = convertFromHumanReadableSize(systemProps.gpuProps.memory)
-    val concurrentGpuTasks: Int = Math.min(
-      gpuMemoryNum * DEFAULT_CONCURRENT_GPU_TASKS_MULTIPLIER,
-      MAX_CONCURRENT_GPU_TASKS).toInt
-    recommendedConfig.setConcurrentGpuTasks(concurrentGpuTasks)
-
-    if(numCores < concurrentGpuTasks) {
-      comments :+= s"For the given GPU, number of CPU cores is very low. It should be" +
-        s" at least equal to concurrent gpu tasks i.e. $concurrentGpuTasks."
-    }
-  }
-
-  /**
-   * Recommendations for Spark only properties.
-   */
-  private def recommendSparkProperties(recommendedConfig: Config,
-      systemProps: SystemProps): Unit = {
-    if (systemProps == null) {
-      logWarning("System information is not available. Cannot recommend properties.")
-      comments :+= "'spark.executor.memory' should be set to at least 2GB/core."
-      comments :+= "'spark.executor.instances' should be set to 'num_gpus * num_workers'."
+  def getRecommendedProperties(
+      skipList: Option[Seq[String]] = None, showOnlyUpdatedProps: Boolean = true):
+      (Seq[RecommendedPropertyResult], Seq[RecommendedCommentResult]) = {
+    filterByUpdatedPropertiesEnabled = showOnlyUpdatedProps
+    skipRecommendations = skipList
+    if (processPropsAndCheck) {
+      initRecommendations()
+      calculateRecommendations()
     } else {
-      recommendExecutorInstances(recommendedConfig, systemProps)
-      recommendExecutorCores(recommendedConfig, systemProps)
-      recommendExecutorMemory(recommendedConfig, systemProps)
-      recommendMemoryOverheadProperties(recommendedConfig)
-      recommendShufflePartitions(recommendedConfig)
-      recommendMaxPartitionBytes(recommendedConfig)
-      recommendGeneralProperties()
+      // add all default comments
+      commentsForMissingProps.foreach(commentEntry => appendComment(commentEntry._2))
     }
-  }
-
-  /**
-   * Recommendations for GPU (RAPIDS) properties.
-   */
-  private def recommendGpuProperties(recommendedConfig: Config, systemProps: SystemProps): Unit = {
-    if (systemProps == null || systemProps.gpuProps == null) {
-      logWarning("GPU information is not available. Cannot recommend properties.")
-      comments :+= "'spark.task.resource.gpu.amount' should be set to 1/#cores."
-      comments :+= s"'spark.rapids.sql.concurrentGpuTasks' should be set to" +
-        s" $DEFAULT_CONCURRENT_GPU_TASKS."
-      comments :+= s"'spark.rapids.memory.pinnedPool.size' should be set to" +
-        s" $DEFAULT_PINNED_POOL_SIZE."
-    } else {
-      recommendTaskResourceGpu(recommendedConfig)
-      recommendConcurrentGpuTasks(recommendedConfig, systemProps)
-    }
-  }
-
-  /**
-   * Entry point for generating recommendations.
-   */
-  def getRecommendedProperties: (Seq[RecommendedPropertyResult],
-    Seq[RecommendedCommentResult]) = {
-    val (systemProps, message) = parseSystemInfo(workerInfo)
-    if(message.nonEmpty) {
-      comments :+= message.get
-    }
-
-    val recommendedConfig = new Config()
-    recommendSparkProperties(recommendedConfig, systemProps)
-    recommendGpuProperties(recommendedConfig, systemProps)
-    (toPropertyProfileResult(recommendedConfig), toCommentProfileResult(comments))
+    (toRecommendationsProfileResult, toCommentProfileResult)
   }
 }
 
 object AutoTuner extends Logging {
-  val DEFAULT_WORKER_INFO: String = "."
+  // Amount of GPU memory to use per concurrent task in megabytes.
+  // Using a bit less than 8GB here since Dataproc clusters advertise T4s as only having
+  // around 14.75 GB and we want to run with 2 concurrent by default on T4s.
+  val DEF_GPU_MEM_PER_TASK_MB = 7500L
+  // Maximum number of concurrent tasks to run on the GPU
+  val MAX_CONC_GPU_TASKS = 4L
+  // Amount of CPU memory to reserve for system overhead (kernel, buffers, etc.) in megabytes
+  val DEF_SYSTEM_RESERVE_MB: Long = 2 * 1024L
+  // Fraction of the executor JVM heap size that should be additionally reserved
+  // for JVM off-heap overhead (thread stacks, native libraries, etc.)
+  val DEF_HEAP_OVERHEAD_FRACTION = 0.1
+  val MAX_JVM_GCTIME_FRACTION = 0.3
+  // Ideal amount of JVM heap memory to request per CPU core in megabytes
+  val DEF_HEAP_PER_CORE_MB: Long = 2 * 1024L
+  // Maximum amount of pinned memory to use per executor in MB
+  val MAX_PINNED_MEMORY_MB: Long = 4 * 1024L
+  // value in MB
+  val MIN_PARTITION_BYTES_RANGE_MB = 128L
+  // value in MB
+  val MAX_PARTITION_BYTES_RANGE_MB = 256L
+  // value in MB
+  val MAX_PARTITION_BYTES_BOUND_MB: Int = 4 * 1024
+  val MAX_PARTITION_BYTES: String = "512m"
+  val DEF_SHUFFLE_PARTITIONS = "200"
+  val DEF_SHUFFLE_PARTITION_MULTIPLIER: Int = 2
+  // Default CPU cores when missing in cluster properties
+  val DEF_WORKER_CPU_CORES: Int = 16
+  // Default GPU count based on number of cores.
+  // GPU by default is to 1 for each 16 cores
+  val DEF_WORKER_CPU_SORES_PER_GPU = 16
+  val DEFAULT_WORKER_INFO_PATH = "./worker_info.yaml"
   val SUPPORTED_SIZE_UNITS: Seq[String] = Seq("b", "k", "m", "g", "t", "p")
 
-  def loadSystemProperties(fileInput: String): Option[java.util.Map[String, Any]] = {
-    val filePath = new Path(fileInput)
-    val fs = FileSystem.get(filePath.toUri, new Configuration())
-    val yaml = new Yaml()
+  val commentsForMissingProps: mutable.Map[String, String] = mutable.LinkedHashMap[String, String](
+    "spark.executor.memory" ->
+      "'spark.executor.memory' should be set to at least 2GB/core.",
+    "spark.executor.instances" ->
+      "'spark.executor.instances' should be set to (gpuCount * numWorkers).",
+    "spark.task.resource.gpu.amount" ->
+      "'spark.task.resource.gpu.amount' should be set to Max(1, (numCores / gpuCount)).",
+    "spark.rapids.sql.concurrentGpuTasks" ->
+      s"'spark.rapids.sql.concurrentGpuTasks' should be set between [1, $MAX_CONC_GPU_TASKS]",
+    "spark.rapids.memory.pinnedPool.size" ->
+      s"'spark.rapids.memory.pinnedPool.size' should be set up to ${MAX_PINNED_MEMORY_MB}m.",
+    "spark.sql.adaptive.enabled" ->
+      "'spark.sql.adaptive.enabled' should be enabled for better performance.")
+
+  val recommendationsTarget: Seq[String] = Seq[String](
+    "spark.executor.instances",
+    "spark.rapids.sql.enabled",
+    "spark.executor.cores",
+    "spark.executor.memory",
+    "spark.rapids.sql.concurrentGpuTasks",
+    "spark.task.resource.gpu.amount",
+    "spark.sql.shuffle.partitions",
+    "spark.sql.files.maxPartitionBytes",
+    "spark.rapids.memory.pinnedPool.size",
+    "spark.executor.memoryOverhead",
+    "spark.executor.memoryOverheadFactor",
+    "spark.kubernetes.memoryOverheadFactor")
+
+  private def handleException(
+      ex: Exception,
+      appInfo: Option[ApplicationSummaryInfo]): AutoTuner = {
+    logError("Exception: " + ex.getStackTrace.mkString("Array(", ", ", ")"))
+    val tuning = new AutoTuner(new ClusterProperties(), appInfo)
+    val msg = ex match {
+      case cEx: ConstructorException => cEx.getContext
+      case _ => if (ex.getCause != null) ex.getCause.toString else ex.toString
+    }
+    tuning.appendComment(msg)
+    tuning
+  }
+
+  def loadClusterPropertiesFromContent(clusterProps: String): Option[ClusterProperties] = {
+    val representer = new Representer
+    val yamlObjNested = new Yaml(new Constructor(classOf[ClusterProperties]), representer)
+    Option(yamlObjNested.load(clusterProps).asInstanceOf[ClusterProperties])
+  }
+
+  def loadClusterProps(filePath: String): Option[ClusterProperties] = {
+    val path = new Path(filePath)
     var fsIs: FSDataInputStream = null
     try {
-      fsIs = fs.open(filePath)
-      Some(yaml.load(fsIs).asInstanceOf[java.util.Map[String, Any]])
+      val fs = FileSystem.get(path.toUri, new Configuration())
+      fsIs = fs.open(path)
+      val reader = new BufferedReader(new InputStreamReader(fsIs))
+      val fileContent = Stream.continually(reader.readLine()).takeWhile(_ != null).mkString("\n")
+      loadClusterPropertiesFromContent(fileContent)
     } finally {
       if (fsIs != null) {
         fsIs.close()
@@ -498,69 +800,34 @@ object AutoTuner extends Logging {
   }
 
   /**
-   * Parses the yaml file and returns system and gpu properties.
-   * See [[SystemProps]] and [[GpuProps]].
+   * Similar to [[buildAutoTuner]] but it allows constructing the AutoTuner without an
+   * existing file. This can be used in testing.
+   *
+   * @param clusterProps the cluster properties as string.
+   * @param appInfo Optional of the profiling container.
+   * @return a new AutoTuner object.
    */
-  def parseSystemInfo(yamlFile: String): (SystemProps, Option[String]) = {
+  def buildAutoTunerFromProps(
+      clusterProps: String,
+      appInfo: Option[ApplicationSummaryInfo]): AutoTuner = {
     try {
-      val loadedProps = loadSystemProperties(yamlFile)
-      if (loadedProps.isDefined) {
-        val rawProps = loadedProps.get.asScala.toMap.filter { case (_, v) => v != null }
-        val rawSystemProps = rawProps("system").asInstanceOf[java.util.Map[String, Any]]
-          .asScala.toMap.filter { case (_, v) => v != null }
-        if (rawSystemProps.nonEmpty) {
-          val rawGpuProps = rawProps("gpu").asInstanceOf[java.util.Map[String, Any]]
-            .asScala.toMap.filter { case (_, v) => v != null }
-
-          val gpuProps = if (rawGpuProps.nonEmpty) {
-            GpuProps(
-              rawGpuProps("count").toString.toInt,
-              rawGpuProps("memory").toString,
-              rawGpuProps("name").toString)
-          } else {
-            null
-          }
-
-          (SystemProps(
-            rawSystemProps.getOrElse("num_cores", 1).toString.toInt,
-            rawSystemProps.getOrElse("cpu_arch", "").toString,
-            rawSystemProps.getOrElse("memory", "0b").toString,
-            rawSystemProps.getOrElse("free_disk_space", "0b").toString,
-            rawSystemProps.getOrElse("time_zone", "").toString,
-            rawSystemProps.get("num_workers").map(_.toString.toInt),
-            gpuProps), None)
-        } else {
-          (null, Some("System properties was empty"))
-        }
-      }
-      else {
-        (null, Some("System properties was empty"))
-      }
+      val clusterPropsOpt = loadClusterPropertiesFromContent(clusterProps)
+      new AutoTuner(clusterPropsOpt.getOrElse(new ClusterProperties()), appInfo)
     } catch {
-      case e: FileNotFoundException =>
-        logError("Exception: " + e.getStackTrace.mkString("Array(", ", ", ")"))
-        (null, Some("System properties file was not found"))
-      case e: NullPointerException =>
-        logError("Exception: " + e.getStackTrace.mkString("Array(", ", ", ")"))
-        (null, Some("System properties file was not formatted correctly."))
       case e: Exception =>
-        logError("Exception: " + e.getStackTrace.mkString("Array(", ", ", ")"))
-        (null, Some(e.toString))
+        handleException(e, appInfo)
     }
   }
 
-  /**
-   * Returns the value of Spark property from the application summary info.
-   * [[RapidsPropertyProfileResult]] is defined as (key:key, rows: [key, value]).
-   * Returns:
-   * a. If the value is "null" or key is not found: None
-   * b. Else: Some(value)
-   */
-  private def getSparkProperty(app: ApplicationSummaryInfo, property: String): Option[String] = {
-    app.sparkProps.collectFirst {
-      case propertyProfile: RapidsPropertyProfileResult
-        if propertyProfile.key == property && propertyProfile.rows(1) != "null" =>
-        propertyProfile.rows(1)
+  def buildAutoTuner(
+      filePath: String,
+      appInfo: Option[ApplicationSummaryInfo]): AutoTuner = {
+    try {
+      val clusterPropsOpt = loadClusterProps(filePath)
+      new AutoTuner(clusterPropsOpt.getOrElse(new ClusterProperties()), appInfo)
+    } catch {
+      case e: Exception =>
+        handleException(e, appInfo)
     }
   }
 
@@ -576,14 +843,29 @@ object AutoTuner extends Logging {
     (sizeNum * Math.pow(1024, SUPPORTED_SIZE_UNITS.indexOf(sizeUnit))).toLong
   }
 
+  def containsMemoryUnits(size: String): Boolean = {
+    val sizesArr = size.toLowerCase.split("(?=[a-z])")
+    if (sizesArr.length > 1) {
+      SUPPORTED_SIZE_UNITS.contains(sizesArr(1))
+    } else {
+      false
+    }
+  }
+
+  def convertToMB(size: String): Long = {
+    convertFromHumanReadableSize(size) / (1024 * 1024)
+  }
+
   /**
    * Converts size from bytes to human readable.
    * Eg, 4194304 -> "4m", 633554 -> "618.70k".
    */
   def convertToHumanReadableSize(size: Long): String = {
-    if(size < 0) return "0b"
+    if (size < 0L) {
+      return "0b"
+    }
 
-    val unitIndex = (Math.log10(size)/Math.log10(1024)).toInt
+    val unitIndex = (Math.log10(size) / Math.log10(1024)).toInt
     assert(unitIndex < SUPPORTED_SIZE_UNITS.size,
       s"$size is too large to convert to human readable size")
 
@@ -605,16 +887,5 @@ object AutoTuner extends Logging {
     val paddedVersions = version1.split("\\.").zipAll(version2.split("\\."), "0", "0")
     val difference = paddedVersions.find { case (a, b) => a != b }
     difference.fold(0) { case (a, b) => a.toInt - b.toInt }
-  }
-
-  private def toPropertyProfileResult(recommendedConfig:Config): Seq[RecommendedPropertyResult] = {
-    val properties = recommendedConfig.getAllProperties
-    properties.collect {
-      case (property, value) if value != null => RecommendedPropertyResult(property, value)
-    }.toSeq.sortBy(_.property)
-  }
-
-  private def toCommentProfileResult(comments:Seq[String]): Seq[RecommendedCommentResult] = {
-    comments.map(RecommendedCommentResult)
   }
 }
