+++ tools/src/main/scala/com/nvidia/spark/rapids/tool/profiling/ProfileArgs.scala
@@ -93,9 +93,9 @@ Usage: java -cp rapids-4-spark-tools_2.12-<version>.jar:$SPARK_HOME/jars/*
   val workerInfo: ScallopOption[String] =
     opt[String](required = false,
       descr = "File path containing the system information of a worker node. It is assumed " +
-        "that all workers are homogenous. It requires the AutoTuner to be enabled. Default is" +
-        "current directory.",
-      default = Some(AutoTuner.DEFAULT_WORKER_INFO))
+        "that all workers are homogenous. It requires the AutoTuner to be enabled. Default is " +
+        "./worker_info.yaml",
+      default = Some(AutoTuner.DEFAULT_WORKER_INFO_PATH))
 
   validate(filterCriteria) {
     case crit if (crit.endsWith("-newest-filesystem") ||
