+++ tools/src/main/scala/com/nvidia/spark/rapids/tool/profiling/Profiler.scala
@@ -24,7 +24,6 @@ import scala.util.control.NonFatal
 
 import com.nvidia.spark.rapids.ThreadFactoryBuilder
 import com.nvidia.spark.rapids.tool.{EventLogInfo, EventLogPathProcessor}
-import com.nvidia.spark.rapids.tool.profiling.AutoTuning.buildAutoTuning
 import org.apache.hadoop.conf.Configuration
 
 import org.apache.spark.internal.Logging
@@ -311,7 +310,7 @@ class Profiler(hadoopConf: Configuration, appArgs: ProfileArgs) extends Logging 
     val removedBMs = healthCheck.getRemovedBlockManager
     val removedExecutors = healthCheck.getRemovedExecutors
     val unsupportedOps = healthCheck.getPossibleUnsupportedSQLPlan
-   
+
     if (printPlans) {
       CollectInformation.printSQLPlans(apps, outputDir)
     }
@@ -451,28 +450,13 @@ class Profiler(hadoopConf: Configuration, appArgs: ProfileArgs) extends Logging 
         Some("Unsupported SQL Ops"))
 
       if (useAutoTuner) {
-        val workerInfo: String = appArgs.workerInfo.getOrElse(AutoTuner.DEFAULT_WORKER_INFO)
-        val autoTuner: AutoTuning = buildAutoTuning(workerInfo, Option(app))
-        val (properties, comments) = autoTuner.getRecommendedProperties
-        profileOutputWriter.writeText("\n### D. Recommended Configuration ###\n")
-
-        if (properties.nonEmpty) {
-          val propertiesToStr = properties.map(_.toString).reduce(_ + "\n" + _)
-          profileOutputWriter.writeText("\nSpark Properties:\n" + propertiesToStr + "\n")
-        } else {
-          profileOutputWriter.writeText("Cannot recommend properties. See Comments.\n")
-        }
-
-        // Comments are optional
-        if (comments.nonEmpty) {
-          val commentsToStr = comments.map(_.toString).reduce(_ + "\n" + _)
-          profileOutputWriter.writeText("\nComments:\n" + commentsToStr + "\n")
-        }
-      }
-      if (false && useAutoTuner) {
-        val workerInfo: String = appArgs.workerInfo.getOrElse(AutoTuner.DEFAULT_WORKER_INFO)
-        val autoTuner: AutoTuner = new AutoTuner(app, workerInfo)
-        val (properties, comments) = autoTuner.getRecommendedProperties
+        val workerInfoPath = appArgs.workerInfo.getOrElse(AutoTuner.DEFAULT_WORKER_INFO_PATH)
+        val autoTuner: AutoTuner = AutoTuner.buildAutoTuner(workerInfoPath, Some(app))
+        // the autotuner allows skipping some properties
+        // e.g. getRecommendedProperties(Some(Seq("spark.executor.instances"))) skips the
+        // recommendation related to executor instances.
+        val (properties, comments) =
+          autoTuner.getRecommendedProperties()
         profileOutputWriter.writeText("\n### D. Recommended Configuration ###\n")
 
         if (properties.nonEmpty) {
